move validation and runtime out of util package
