+++ pkg/distributor/ingestion_rate_strategy.go
@@ -3,7 +3,7 @@ package distributor
 import (
 	"github.com/cortexproject/cortex/pkg/util/limiter"
 
-	"github.com/grafana/loki/pkg/util/validation"
+	"github.com/grafana/loki/pkg/validation"
 )
 
 // ReadLifecycler represents the read interface to the lifecycler.
