+++ pkg/querier/querier_test.go
@@ -11,6 +11,7 @@ import (
 	ring_client "github.com/cortexproject/cortex/pkg/ring/client"
 
 	"github.com/grafana/loki/pkg/ingester/client"
+	"github.com/grafana/loki/pkg/validation"
 
 	"github.com/grafana/loki/pkg/storage"
 
@@ -27,7 +28,6 @@ import (
 	"github.com/cortexproject/cortex/pkg/util/flagext"
 
 	"github.com/grafana/loki/pkg/logproto"
-	"github.com/grafana/loki/pkg/util/validation"
 )
 
 const (
